Move useS3Upload into hooks folder
